Izmenjen middleware, api

+++ routes/api.php
@@ -6,6 +6,11 @@ use Illuminate\Support\Facades\Route;
 
 use App\Http\Controllers\API\AuthController;
 use App\Http\Controllers\UserController;
+use App\Http\Controllers\DestinacijaController;
+use App\Http\Controllers\HotelController;
+use App\Http\Controllers\PDFExportController;
+use App\Http\Controllers\PlanPutovanjaController;
+use App\Http\Controllers\ZnamenitostController;
 
 /*
 |--------------------------------------------------------------------------
@@ -23,37 +28,36 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-
-
-Route::get('/users', [UserController::class, 'index']);
-Route::get('/users/{user_id}', [UserController::class, 'show']);
-//Route::post('/users/store', [UserController::class, 'store']);
-//Route::put('/users/{id}/update', [UserController::class, 'update']);
-Route::delete('/users/{id}', [UserController::class, 'destroy']);
-Route::resource('users', UserController::class);
-
-Route::post('/register',[AuthController::class, 'register']);
+//Route::resource('destinacije', DestinacijaController::class)->only(['index','show']);
+//Route::resource('znamenitosti', ZnamenitostController::class);
+//Route::resource('hoteli', HotelController::class);
+Route::post('/register', [AuthController::class, 'register']);
 Route::post('/login', [AuthController::class, 'login']);
 
 
-//Route::resource('destinacije', DestinacijaController::class);
-//Route::resource('znamenitosti', ZnamenitostController::class);
-//Route::resource('hoteli', HotelController::class);
+//Route::post('/users/store', [UserController::class, 'store']);
+//Route::put('/users/{id}/update', [UserController::class, 'update']);
+//Route::delete('/users/{id}', [UserController::class, 'destroy']);
 
-/*Route::group(['middleware'=>['auth:sanctum','role:ulogovan']], function(){
-    Route::get('/user/profile',function(Request $request){
-        return $request->user();
-    });
-
-    Route::post('/logout', [AuthController::class, 'logout']);
-});*/
-
+//samo za ulogovane
 Route::group(['middleware' => ['auth:sanctum']], function () {
     Route::get('/profile', function(Request $request) {
         return auth()->user();
     });
+    
+   // Route::get('/user', [UserController::class, 'show']);
+   Route::get('/users', [UserController::class, 'index'])->middleware('accessControl:1');
+   Route::get('/users/{id}', [UserController::class, 'show'])->middleware('accessControl:1');
+   Route::delete('/users/{id}', [UserController::class, 'destroy'])->middleware('accessControl:1');
 
-    // API route for logout user
-    Route::post('/logout', [AuthController::class, 'logout']);
+   Route::get('/planPutovanja', [PlanPutovanjaController::class, 'index'])->middleware('accessControl:1');
+
+   //Login user moze da kreira, izmenjuje i brise svoja putovanja
+   Route::post('/storePlanPutovanja', [PlanPutovanjaController::class, 'store'])->middleware('accessControl:2');
+   Route::put('/updatePlanPutovanja/{id}', [PlanPutovanjaController::class, 'update'])->middleware('accessControl:2');
+   Route::delete('/deletePlanPutovanja/{id}', [PlanPutovanjaController::class, 'destroy'])->middleware('accessControl:2');
+
+   Route::post('/logout', [AuthController::class, 'logout']);
+
 });
 Route::post('forgot/password',[AuthController::class,'forgotPassword']);
